backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
langchain==0.0.195
langchain-community==0.0.28
langchain-openai==0.0.8
pypdf==4.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.28
pydantic==2.6.4
python-jose==3.3.0
openai==1.14.1
faiss-cpu==1.8.0

